chore: refactored query for readability
